Fixes controllers
