+++ app/controllers/admin_script/application_controller.rb
@@ -1,0 +1,5 @@
+module AdminScript
+  # TODO: Change parent class
+  class ApplicationController < ActionController::Base
+  end
+end
