--- app/views/admin_scripts/_default_form.html.slim
@@ -1,18 +1,0 @@
-- f.object.class.type_attributes.each do |name, type|
-  .form-group
-    = f.label name
-    - case type
-    - when :integer, :big_integer
-      = f.number_field name, class: 'form-control'
-    - when :string, :float
-      = f.text_field name, class: 'form-control'
-    - when :date
-      = f.date_field name, class: 'form-control'
-    - when :date_time
-      = f.datetime_select name, class: 'form-control'
-    - when :binary
-      = f.file_field name, class: 'form-control'
-    - when :text
-      = f.text_field name, class: 'form-control'
-    - when :time
-      = f.time_select name, class: 'form-control'
