--- app/views/admin_scripts/show.html.slim
@@ -1,19 +1,0 @@
-h1.page-header AdminScript #{@admin_script.id}
-
-.panel.panel-default
-  .panel-body
-    = form_for @admin_script, url: admin_script_path(@admin_script), as: :admin_script do |f|
-      .form-group
-        = f.label :description
-        p = simple_format f.object.description
-
-      - template = "admin_scripts/#{@admin_script.model_name.element}"
-      - if lookup_context.template_exists?(template, controller._prefixes, true)
-        = render @admin_script.model_name.element, f: f
-      - else
-        = render 'admin_scripts/default_form', f: f
-
-      = f.button 'Perform', class: 'btn btn-default', data: { confirm: 'Are you sure?' }
-
-  .panel-footer
-    pre = @admin_script.class.script
